py-harvester/pyproject.toml
[tool.poetry]
name = "py-harvester"
version = "0.1.0"
description = ""
authors = ["king-ak"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
marshmallow = "^3.12.1"
simplejson = "^3.17.5"
backoff = "^2.2.1"
kafka-python = "^2.0.2"

[tool.poetry.group.test.dependencies]
pytest = "^6.0.1"
parameterized = "^0.7.4"
flake8 = "^3.8.4"
autopep8 = "^1.4.4"


##
## Postgres connection
#psycopg2-binary==2.8.5
## Alembic/SQLAlchemy stuff
#SQLAlchemy==1.4.20
#alembic==1.6.5
#Werkzeug==2.0.1
#marshmallow==3.12.1
#marshmallow-sqlalchemy==0.26.1
#simplejson==3.17.5
#
## Stuff to pull external data
#lxml==4.5.2
#requests==2.25.1
#yahoo-fin==0.8.9.1
#html5lib==1.1
#cryptocompare==0.7.5
#backoff==1.11.1
#
#
## Kafka
#kafka-python==2.0.2
#
## Analytics/Evaluation
## matplotlib==3.3.4
#pandas==1.3.1
#panel==0.12.6
#pytest==6.0.1
#parameterized==0.7.4
## Linting/formatting
#flake8==3.8.4
#autopep8==1.4.4

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
